backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.2
pydantic[email]
pydantic-settings>=2.0.0,<3.0.0
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator>=2.0.0,<3.0.0
python-dotenv==1.0.0
openai>=1.0.0,<2.0.0
httpx>=0.24.0
authlib>=1.2.0
requests>=2.31.0
pyotp>=2.8.0
qrcode>=7.4.2
pillow>=10.0.0
redis>=5.0.0,<6.0.0
motor>=3.0.0
beanie>=1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-env==1.0.1
pytest-xdist==3.3.1

# Development dependencies
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Monitoring and Logging
prometheus-client==0.17.1
# sentry-sdk>=1.40.0,<2.0.0
python-json-logger==2.0.7
structlog==23.1.0

# Password strength and breach detection
zxcvbn==4.4.28

# Email
fastapi-mail>=1.4.1,<1.5.0
aiosmtplib>=2.0.2
jinja2>=3.1.2

# File handling
aiofiles>=23.2.1

# Security
cryptography>=41.0.7,<42.0.0
bleach>=6.1.0,<7.0.0
requests_oauthlib>=1.3.1,<2.0.0

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-env==1.0.1
pytest-xdist==3.3.1

# Development dependencies
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Monitoring and Logging
prometheus-client==0.17.1
# sentry-sdk>=1.40.0,<2.0.0
python-json-logger==2.0.7
structlog==23.1.0

# Password strength and breach detection
zxcvbn==4.4.28

# Email
fastapi-mail>=1.4.1,<1.5.0
aiosmtplib>=2.0.2
jinja2>=3.1.2

# File handling
aiofiles>=23.2.1 